% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\name{AIC_adapted}
\alias{AIC_adapted}
\title{Calculate the adapted AIC score of a set of estimated precision matrices}
\usage{
AIC_adapted(theta, sample.cov, n.vals)
}
\arguments{
\item{theta}{The list of estimated precision matrices.}

\item{sample.cov}{The list of sample covariance matrices of the different sets of observed data.}

\item{n.vals}{The number of observations in each class. A vector.}
}
\value{
The adapted AIC score.
}
\description{
Finds the adapted AIC score of an estimated precision matrix given a sample covariance matrix, in the setting of Gaussian graphical model.
}
\examples{

# example 1: simple example where we check the adapted AIC score
#            of a set of the true precision matrices
# generate data
set.seed(123)
n <- 80
p <- 10
dat1 <- huge::huge.generator(n = n, d = p, graph = "scale-free")
dat2 <- huge::huge.generator(n = n, d = p, graph = "scale-free")
cov.list = list(var(dat1$data), var(dat2$data))
prec.mat1 <- dat1$omega # true precision matrix
prec.mat2 <- dat2$omega # true precision matrix
prec.mat.list = list(prec.mat1,prec.mat2)
AIC_adapted(cov.list, prec.mat.list, n.vals=c(n,n))
}
\seealso{
\code{\link{JoStARS}}
}
\author{
Camilla Lingjaerde
}
