% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\name{eBIC_adapted}
\alias{eBIC_adapted}
\title{Calculate the adapted extended BIC (adapted eBIC) score of a set of estimated precision matrices}
\usage{
eBIC_adapted(theta, sample.cov, n.vals, gamma = 0.2)
}
\arguments{
\item{theta}{The list of estimated precision matrices.}

\item{sample.cov}{The list of sample covariance matrices of the different sets of observed data.}

\item{n.vals}{The number of observations in each class. A vector.}

\item{gamma}{The value of the additional edge penalty parameter \eqn{\gamma} to use in the adapted eBIC. Default value is \eqn{0.2}.}
}
\value{
The adapted eBIC score.
}
\description{
Finds the adapted extended BIC (adapted eBIC) score of an estimated precision matrix given a sample covariance matrix, in the setting of Gaussian graphical model.
}
\examples{

# example 1: simple example where we check the adapted eBIC
#            score of a set of the true precision matrices
# generate data
set.seed(123)
n <- 80
p <- 10
dat1 <- huge::huge.generator(n = n, d = p, graph = "scale-free")
dat2 <- huge::huge.generator(n = n, d = p, graph = "scale-free")
cov.list = list(var(dat1$data), var(dat2$data))
prec.mat1 <- dat1$omega # true precision matrix
prec.mat2 <- dat2$omega # true precision matrix
prec.mat.list = list(prec.mat1,prec.mat2)
eBIC_adapted(cov.list, prec.mat.list, n.vals=c(n,n), gamma = 0.2)

}
\seealso{
\code{\link{stabJGL}}
}
\author{
Camilla Lingjaerde
}
