% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoStARS.R
\name{JoStARS_select_lambda1_parallel}
\alias{JoStARS_select_lambda1_parallel}
\title{Internal function used when selecting lambda1 with several threads in JoStARS}
\usage{
JoStARS_select_lambda1_parallel(
  Y,
  rep.num,
  rho,
  n.vals,
  stars.subsample.ratios,
  lambda1s,
  lambda2,
  penalize.diagonal,
  seed,
  array.list
)
}
\description{
Internal function used when selecting lambda1 with several threads in JoStARS
}
\keyword{internal}
