% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoStARS.R
\name{JoStARS_select_lambda2_eBIC}
\alias{JoStARS_select_lambda2_eBIC}
\title{Internal function for selecting lambda2 in JoStARS}
\usage{
JoStARS_select_lambda2_eBIC(
  Y,
  penalty = "fused",
  rho = 1,
  weights = "equal",
  penalize.diagonal = FALSE,
  nlambda2 = 30,
  lambda2.min,
  lambda2.max,
  lambda1 = NULL,
  gamma = ebic.gamma,
  verbose,
  parallelize = F,
  nCores = 4
)
}
\description{
Internal function for selecting lambda2 in JoStARS
}
\keyword{internal}
