% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabJGL.R
\name{stabJGL_select_lambda1}
\alias{stabJGL_select_lambda1}
\title{Internal function for selecting lambda1 in stabJGL}
\usage{
stabJGL_select_lambda1(
  Y,
  rho = 1,
  weights = "equal",
  penalize.diagonal = FALSE,
  stars.thresh = 0.1,
  stars.subsample.ratio = NULL,
  rep.num = 20,
  nlambda1 = 20,
  lambda1.min,
  lambda1.max,
  lambda2,
  verbose,
  parallelize = F,
  nCores = 4
)
}
\description{
Internal function for selecting lambda1 in stabJGL
}
\keyword{internal}
