% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabJGL.R
\name{stabJGL_select_lambda2_eBIC}
\alias{stabJGL_select_lambda2_eBIC}
\title{Internal function for selecting lambda2 in stabJGL}
\usage{
stabJGL_select_lambda2_eBIC(
  Y,
  rho = 1,
  weights = "equal",
  penalize.diagonal = FALSE,
  nlambda2 = 30,
  lambda2.min,
  lambda2.max,
  lambda1 = NULL,
  gamma = NULL,
  verbose = F,
  parallelize = F,
  nCores = 4
)
}
\description{
Internal function for selecting lambda2 in stabJGL
}
\keyword{internal}
